Moved classes to independent files
